Algoritmo 2 terminado

+++ src/utilidades/Punto.java
@@ -2,28 +2,28 @@ package utilidades;
 
 public class Punto {
 
-	int x;
-	int y;
+	double x;
+	double y;
 	
 	
 
-	public Punto(int x, int y) {
+	public Punto(double x, double y) {
 		super();
 		this.x = x;
 		this.y = y;
 	}
 	
 	
-	public int getX() {
-		return x;
+	public double getX() {
+		return  x;
 	}
-	public void setX(int x) {
+	public void setX(Double x) {
 		this.x = x;
 	}
-	public int getY() {
+	public double getY() {
 		return y;
 	}
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
